requirements.txt
requests>=2.26.0
matplotlib>=3.4.3
pyvis>=0.2.1
python-dotenv>=0.19.0
networkx>=2.6.3